requirements.txt
Flask==3.0.3
flask-cors==4.0.0
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0
numpy==2.1.1
scikit-learn==1.5.2
joblib==1.4.2
boto3==1.35.14
botocore==1.35.14
tqdm==4.66.5
openpyxl>=3.1.5
pandas>=2.2.3


